# Optimize prefill performance on cpu backend