## Motivation

This PR aims at improving prefill performance for CPU backend, it will improve `bfloat16`, `int8_w8a8` and `fp8_w8a8` paths. This one has no effective on decoding performance.

## Modifications

* **amx-int8**: enable **amx-int8** for GEMM and MoE kernels. As SGLang pops up with PyTorch 2.7, we are free to use amx-int8 on SGLang CPU backend.
* **thread blocking**: the original code base utilizes a simple blocking scheme for parallel, due to the limited time frame allowed for this project. Now i added more complexed blocking scheme. to use `parallel_2d` to replace `parallel_for` (which a 1d parallel template). Which means that for a 2d problem size of [M, N], the available threads will also be chunked to `num_threads_m * num_threads_n`, where the problem allocated to each thread will be as square as possible - so as to minimize memory load globally.
* **cache blocking**: use `loop_2d` template to alter the loop order from [MB, NB] to [NB / cache_blocks_nb, MB, cache_blocks_nb] -  so as to maximize the usage of L2 cache as much as possible.
* **fp8 dequant blocking**: fp8 dequant is super slow on CPU, so here I manage to dequant only once per thread and cache the bf16 weights in L2 and reuse it as much as possible.


## Accuracy Test

we can test with:
```bash
python -m unittest -v test_gemm.py
python -m unittest -v test_moe.py
```

this PR is purely backend performance improvement and has no effect on end2end accuracy.

## Benchmark & Profiling

i tested the PR on Intel 4th gen Xeon, (codename SPR), single socket with 40 cores. (the optimization in this PR also applies to 5th and 6th gen Xeon, e.g. EMR and GNR).

to reproduce the performance, use [sgl-cpu-test](https://github.com/mingfeima/sgl-cpu-tests)

```bash
./run_bench_cpu.sh bench_gemm.py
./run_bench_cpu.sh bench_moe.py
```

### Performance Improvement of GEMM
```
### before
### gemm_fp8 benchmark: M = 1024, N = 14336, K = 4096, has_bias = False
gemm_bf16(native): 4.644 ms, gemm_fp8(opt): 4.375 ms, gemm_int8(opt): 8.945 ms, gemm_bf16(opt): 14.843 ms

### after
### gemm_fp8 benchmark: M = 1024, N = 14336, K = 4096, has_bias = False
gemm_bf16(native): 4.805 ms, gemm_fp8(opt): 3.483 ms, gemm_int8(opt): 2.048 ms, gemm_bf16(opt): 3.975 ms
```

### Performance Improvement of MoE
```
### before
### fused_experts: M = 4, N = 384, K = 7168, E = 256, TopK = 8: bfloat16: 2.124 ms; int8: 1.057 ms; fp8: 1.066 ms
### fused_experts: M = 3929, N = 384, K = 7168, E = 256, TopK = 8: bfloat16: 73.836 ms; int8: 53.745 ms; fp8: 60.054 ms

### after
### fused_experts: M = 4, N = 384, K = 7168, E = 256, TopK = 8: bfloat16: 2.085 ms; int8: 1.081 ms; fp8: 1.100 ms
### fused_experts: M = 3929, N = 384, K = 7168, E = 256, TopK = 8: bfloat16: 38.431 ms; int8: 23.258 ms; fp8: 44.205 ms
```

Notes:
* shapes collected from deepseek r1 with TP 6.
* int8 flops should be pretty much 2x of bfloat16.
* fp8 flops as close to bfloat16 as possible. 

## Checklist

- [x] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [x] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.