@mingfeima 2025-08-05
waiting our validation team provide model level performance/accuracy results.