@Alcanderian 2025-08-25
Do we have any plans to optimize the brgemm-scaleC fusion?