@mingfeima 2025-08-25
Our validation team has finished performance data collection, and in general:

* FP8 prefill throughput increased by **1.2x**
* INT8 prefill throughput increased by **2.4x**
* BF16 prefill throughput increased by **1.6x**

NB: FP8 smaller speedup as current x86 does not have native instructions for FP8 and the data type conversion from FP8 to BF16 is super slow.


(Detailed data listed below for input/output sequence with 1024/1024, measured on Intel Xeon )

#### FP8

Model | DTYPE | TP | Concurrency | TTFT (ms) Before | TTFT (ms) After | Speedup
-- | -- | -- | -- | -- | -- | --
Qwen--Qwen3-14B-FP8 | FP8 | 1 | 1 | 1514.03 | 1693.2 | 111.83%
Qwen--Qwen3-14B-FP8 | FP8 | 1 | 2 | 2271.07 | 2536.62 | 111.69%
Qwen--Qwen3-14B-FP8 | FP8 | 1 | 4 | 4661.82 | 5745.72 | 123.25%
Qwen--Qwen3-14B-FP8 | FP8 | 1 | 8 | 8306.86 | 10386.55 | 125.04%
Qwen--Qwen3-14B-FP8 | FP8 | 1 | 16 | 14302.84 | 18075.56 | 126.38%
Qwen--Qwen3-14B-FP8 | FP8 | 2 | 1 | 726.63 | 938.34 | 129.14%
Qwen--Qwen3-14B-FP8 | FP8 | 2 | 2 | 1088.92 | 1392.14 | 127.85%
Qwen--Qwen3-14B-FP8 | FP8 | 2 | 4 | 1963.99 | 2858.12 | 145.53%
Qwen--Qwen3-14B-FP8 | FP8 | 2 | 8 | 3894.91 | 5053.86 | 129.76%
Qwen--Qwen3-14B-FP8 | FP8 | 2 | 16 | 6584.01 | 8663.48 | 131.58%
Qwen--Qwen3-14B-FP8 | FP8 | 3 | 1 | 582.34 | 704.77 | 121.02%
Qwen--Qwen3-14B-FP8 | FP8 | 3 | 2 | 1037.74 | 1050.81 | 101.26%
Qwen--Qwen3-14B-FP8 | FP8 | 3 | 4 | 1523.51 | 2078.28 | 136.41%
Qwen--Qwen3-14B-FP8 | FP8 | 3 | 8 | 2944.3 | 3581.94 | 121.66%
Qwen--Qwen3-14B-FP8 | FP8 | 3 | 16 | 5120.78 | 6133.21 | 119.77%
Qwen--Qwen3-235B-A22B-Instruct-2507-FP8 | FP8 | 6 | 1 | 1255.95 | 1350.41 | 107.52%
Qwen--Qwen3-235B-A22B-Instruct-2507-FP8 | FP8 | 6 | 2 | 1879.5 | 2027.89 | 107.90%
Qwen--Qwen3-235B-A22B-Instruct-2507-FP8 | FP8 | 6 | 4 | 3513.53 | 3883.5 | 110.53%
Qwen--Qwen3-235B-A22B-Instruct-2507-FP8 | FP8 | 6 | 8 | 6220.38 | 6867.17 | 110.40%
Qwen--Qwen3-235B-A22B-Instruct-2507-FP8 | FP8 | 6 | 16 | 10562.43 | 11672.79 | 110.51%
deepseek-ai--DeepSeek-R1 | FP8 | 6 | 1 | 2034.16 | 2249.89 | 110.61%
deepseek-ai--DeepSeek-R1 | FP8 | 6 | 2 | 3013.4 | 3415 | 113.33%
deepseek-ai--DeepSeek-R1 | FP8 | 6 | 4 | 5480.19 | 6428.35 | 117.30%
deepseek-ai--DeepSeek-R1 | FP8 | 6 | 8 | 9760.98 | 12033.56 | 123.28%
deepseek-ai--DeepSeek-R1 | FP8 | 6 | 16 | 17421.65 | 21264.43 | 122.06%


#### INT8

Model | DTYPE | TP | Concurrency | TTFT (ms) Before | TTFT (ms) After | Speedup
-- | -- | -- | -- | -- | -- | --
RedHatAI--Llama-3.2-3B-quantized.w8a8 | INT8 | 1 | 1 | 220.68 | 570.74 | 258.63%
RedHatAI--Llama-3.2-3B-quantized.w8a8 | INT8 | 1 | 2 | 338.1 | 855.83 | 253.13%
RedHatAI--Llama-3.2-3B-quantized.w8a8 | INT8 | 1 | 4 | 638.2 | 1647.81 | 258.20%
RedHatAI--Llama-3.2-3B-quantized.w8a8 | INT8 | 1 | 8 | 1107.98 | 2911.8 | 262.80%
RedHatAI--Llama-3.2-3B-quantized.w8a8 | INT8 | 1 | 16 | 1893.37 | 5000.87 | 264.13%
RedHatAI--Llama-3.2-3B-quantized.w8a8 | INT8 | 2 | 1 | 141.46 | 306.45 | 216.63%
RedHatAI--Llama-3.2-3B-quantized.w8a8 | INT8 | 2 | 2 | 255.15 | 542.36 | 212.57%
RedHatAI--Llama-3.2-3B-quantized.w8a8 | INT8 | 2 | 4 | 378.17 | 859.87 | 227.38%
RedHatAI--Llama-3.2-3B-quantized.w8a8 | INT8 | 2 | 8 | 712.77 | 1551.21 | 217.63%
RedHatAI--Llama-3.2-3B-quantized.w8a8 | INT8 | 2 | 16 | 1239.61 | 2628.9 | 212.07%
RedHatAI--Llama-3.2-3B-quantized.w8a8 | INT8 | 3 | 1 | 127.02 | 235.42 | 185.34%
RedHatAI--Llama-3.2-3B-quantized.w8a8 | INT8 | 3 | 2 | 180.14 | 399.78 | 221.93%
RedHatAI--Llama-3.2-3B-quantized.w8a8 | INT8 | 3 | 4 | 332.49 | 657.24 | 197.67%
RedHatAI--Llama-3.2-3B-quantized.w8a8 | INT8 | 3 | 8 | 530.12 | 1233.79 | 232.74%
RedHatAI--Llama-3.2-3B-quantized.w8a8 | INT8 | 3 | 16 | 947.82 | 2060.09 | 217.35%
RedHatAI--Meta-Llama-3.1-8B-quantized.w8a8 | INT8 | 1 | 1 | 447.05 | 1265.04 | 282.98%
RedHatAI--Meta-Llama-3.1-8B-quantized.w8a8 | INT8 | 1 | 2 | 677.19 | 1899.19 | 280.45%
RedHatAI--Meta-Llama-3.1-8B-quantized.w8a8 | INT8 | 1 | 4 | 1285.39 | 3698.63 | 287.74%
RedHatAI--Meta-Llama-3.1-8B-quantized.w8a8 | INT8 | 1 | 8 | 2224.03 | 6671 | 299.95%
RedHatAI--Meta-Llama-3.1-8B-quantized.w8a8 | INT8 | 1 | 16 | 3812.69 | 11594.14 | 304.09%
RedHatAI--Meta-Llama-3.1-8B-quantized.w8a8 | INT8 | 2 | 1 | 284.15 | 726.8 | 255.78%
RedHatAI--Meta-Llama-3.1-8B-quantized.w8a8 | INT8 | 2 | 2 | 487.66 | 1240.79 | 254.44%
RedHatAI--Meta-Llama-3.1-8B-quantized.w8a8 | INT8 | 2 | 4 | 773.12 | 2021.56 | 261.48%
RedHatAI--Meta-Llama-3.1-8B-quantized.w8a8 | INT8 | 2 | 8 | 1280.54 | 3600.87 | 281.20%
RedHatAI--Meta-Llama-3.1-8B-quantized.w8a8 | INT8 | 2 | 16 | 2266.4 | 6176.5 | 272.52%
RedHatAI--Meta-Llama-3.1-8B-quantized.w8a8 | INT8 | 3 | 1 | 217.08 | 515.29 | 237.37%
RedHatAI--Meta-Llama-3.1-8B-quantized.w8a8 | INT8 | 3 | 2 | 379.04 | 779.28 | 205.59%
RedHatAI--Meta-Llama-3.1-8B-quantized.w8a8 | INT8 | 3 | 4 | 684.45 | 1510.26 | 220.65%
RedHatAI--Meta-Llama-3.1-8B-quantized.w8a8 | INT8 | 3 | 8 | 1052.52 | 2828.34 | 268.72%
RedHatAI--Meta-Llama-3.1-8B-quantized.w8a8 | INT8 | 3 | 16 | 1795.84 | 4568.88 | 254.41%
RedHatAI--QwQ-32B-quantized.w8a8 | INT8 | 2 | 1 | 980.38 | 2710.28 | 276.45%
RedHatAI--QwQ-32B-quantized.w8a8 | INT8 | 2 | 2 | 1461.51 | 5113.12 | 349.85%
RedHatAI--QwQ-32B-quantized.w8a8 | INT8 | 2 | 4 | 2737.21 | 7634.59 | 278.92%
RedHatAI--QwQ-32B-quantized.w8a8 | INT8 | 2 | 8 | 4687.5 | 14896.75 | 317.80%
RedHatAI--QwQ-32B-quantized.w8a8 | INT8 | 2 | 16 | 7911.37 | 25889.72 | 327.25%
RedHatAI--QwQ-32B-quantized.w8a8 | INT8 | 3 | 1 | 734.17 | 2041.95 | 278.13%
RedHatAI--QwQ-32B-quantized.w8a8 | INT8 | 3 | 2 | 1098.05 | 3165.81 | 288.31%
RedHatAI--QwQ-32B-quantized.w8a8 | INT8 | 3 | 4 | 1841.89 | 5709.14 | 309.96%
RedHatAI--QwQ-32B-quantized.w8a8 | INT8 | 3 | 8 | 3150.44 | 10761.49 | 341.59%
RedHatAI--QwQ-32B-quantized.w8a8 | INT8 | 3 | 16 | 5787.7 | 19244.92 | 332.51%
RedHatAI--DeepSeek-R1-Distill-Llama-70B-quantized.w8a8 | INT8 | 6 | 1 | 1055.11 | 2609.55 | 247.32%
RedHatAI--DeepSeek-R1-Distill-Llama-70B-quantized.w8a8 | INT8 | 6 | 2 | 1568.73 | 3993.35 | 254.56%
RedHatAI--DeepSeek-R1-Distill-Llama-70B-quantized.w8a8 | INT8 | 6 | 4 | 3112.77 | 8010.18 | 257.33%
RedHatAI--DeepSeek-R1-Distill-Llama-70B-quantized.w8a8 | INT8 | 6 | 8 | 4613.02 | 14684.05 | 318.32%
RedHatAI--DeepSeek-R1-Distill-Llama-70B-quantized.w8a8 | INT8 | 6 | 16 | 8278.83 | 25171.69 | 304.05%
Qwen3-235B-A22B-Instruct-2507-INT8 | INT8 | 6 | 1 | 832.33 | 1181.82 | 141.99%
Qwen3-235B-A22B-Instruct-2507-INT8 | INT8 | 6 | 2 | 1263.7 | 2151.08 | 170.22%
Qwen3-235B-A22B-Instruct-2507-INT8 | INT8 | 6 | 4 | 2296.43 | 3488.48 | 151.91%
Qwen3-235B-A22B-Instruct-2507-INT8 | INT8 | 6 | 8 | 4003.21 | 6292.64 | 157.19%
Qwen3-235B-A22B-Instruct-2507-INT8 | INT8 | 6 | 16 | 6791.38 | 10819.99 | 159.32%
**meituan--DeepSeek-R1-Channel-INT8** | INT8 | 6 | 1 | 1231.54 | 1925.04 | 156.31%
**meituan--DeepSeek-R1-Channel-INT8** | INT8 | 6 | 2 | 1862 | 2964.34 | 159.20%
**meituan--DeepSeek-R1-Channel-INT8** | INT8 | 6 | 4 | 3330.21 | 5784.32 | 173.69%
**meituan--DeepSeek-R1-Channel-INT8** | INT8 | 6 | 8 | 6062.73 | 10673.97 | 176.06%
**meituan--DeepSeek-R1-Channel-INT8** | INT8 | 6 | 16 | 10824.88 | 19106.06 | 176.50%


#### BF16

Model | DTYPE | TP | Concurrency | TTFT (ms) Before | TTFT (ms) After | Speedup
-- | -- | -- | -- | -- | -- | --
meta-llama--Llama-3.2-3B-Instruct | bf16 | 1 | 1 | 371.64 | 578.2 | 155.58%
meta-llama--Llama-3.2-3B-Instruct | bf16 | 1 | 2 | 561.71 | 865.65 | 154.11%
meta-llama--Llama-3.2-3B-Instruct | bf16 | 1 | 4 | 1031.29 | 1445.83 | 140.20%
meta-llama--Llama-3.2-3B-Instruct | bf16 | 1 | 8 | 1801.84 | 2460.52 | 136.56%
meta-llama--Llama-3.2-3B-Instruct | bf16 | 1 | 16 | 3072.9 | 4126.36 | 134.28%
meta-llama--Llama-3.2-3B-Instruct | bf16 | 2 | 1 | 215.88 | 338.22 | 156.67%
meta-llama--Llama-3.2-3B-Instruct | bf16 | 2 | 2 | 328.76 | 513.19 | 156.10%
meta-llama--Llama-3.2-3B-Instruct | bf16 | 2 | 4 | 587.21 | 809.13 | 137.79%
meta-llama--Llama-3.2-3B-Instruct | bf16 | 2 | 8 | 1013.04 | 1393.82 | 137.59%
meta-llama--Llama-3.2-3B-Instruct | bf16 | 2 | 16 | 1720.41 | 2343.97 | 136.24%
meta-llama--Llama-3.2-3B-Instruct | bf16 | 3 | 1 | 172.99 | 266.21 | 153.89%
meta-llama--Llama-3.2-3B-Instruct | bf16 | 3 | 2 | 256.64 | 400.64 | 156.11%
meta-llama--Llama-3.2-3B-Instruct | bf16 | 3 | 4 | 461.29 | 655.16 | 142.03%
meta-llama--Llama-3.2-3B-Instruct | bf16 | 3 | 8 | 738.21 | 1101.03 | 149.15%
meta-llama--Llama-3.2-3B-Instruct | bf16 | 3 | 16 | 1322.71 | 1789.64 | 135.30%
meta-llama--Llama-3.1-8B-Instruct | bf16 | 1 | 1 | 804.22 | 1374.58 | 170.92%
meta-llama--Llama-3.1-8B-Instruct | bf16 | 1 | 2 | 1206.09 | 2062.6 | 171.02%
meta-llama--Llama-3.1-8B-Instruct | bf16 | 1 | 4 | 2308.23 | 3837.21 | 166.24%
meta-llama--Llama-3.1-8B-Instruct | bf16 | 1 | 8 | 4037.58 | 6807 | 168.59%
meta-llama--Llama-3.1-8B-Instruct | bf16 | 1 | 16 | 6911.32 | 11602.02 | 167.87%
meta-llama--Llama-3.1-8B-Instruct | bf16 | 2 | 1 | 443.06 | 732.63 | 165.36%
meta-llama--Llama-3.1-8B-Instruct | bf16 | 2 | 2 | 653.84 | 1092.75 | 167.13%
meta-llama--Llama-3.1-8B-Instruct | bf16 | 2 | 4 | 1325.1 | 1740.53 | 131.35%
meta-llama--Llama-3.1-8B-Instruct | bf16 | 2 | 8 | 2079.62 | 2895.13 | 139.21%
meta-llama--Llama-3.1-8B-Instruct | bf16 | 2 | 16 | 3522 | 4952.93 | 140.63%
meta-llama--Llama-3.1-8B-Instruct | bf16 | 3 | 1 | 357.67 | 549.74 | 153.70%
meta-llama--Llama-3.1-8B-Instruct | bf16 | 3 | 2 | 532.31 | 843.74 | 158.51%
meta-llama--Llama-3.1-8B-Instruct | bf16 | 3 | 4 | 866.74 | 1211.11 | 139.73%
meta-llama--Llama-3.1-8B-Instruct | bf16 | 3 | 8 | 1602.1 | 2301.54 | 143.66%
meta-llama--Llama-3.1-8B-Instruct | bf16 | 3 | 16 | 2720.77 | 3907.75 | 143.63%
Qwen--Qwen3-14B | bf16 | 1 | 1 | 1511.45 | 2987.11 | 197.63%
Qwen--Qwen3-14B | bf16 | 1 | 2 | 2268.66 | 4478.06 | 197.39%
Qwen--Qwen3-14B | bf16 | 1 | 4 | 4523.96 | 8729.65 | 192.96%
Qwen--Qwen3-14B | bf16 | 1 | 8 | 7987.74 | 14635.78 | 183.23%
Qwen--Qwen3-14B | bf16 | 1 | 16 | 13658.46 | 24451.5 | 179.02%
Qwen--Qwen3-14B | bf16 | 2 | 1 | 786.22 | 1241.71 | 157.93%
Qwen--Qwen3-14B | bf16 | 2 | 2 | 1353.29 | 2081.72 | 153.83%
Qwen--Qwen3-14B | bf16 | 2 | 4 | 2227.09 | 3086.47 | 138.59%
Qwen--Qwen3-14B | bf16 | 2 | 8 | 3895.06 | 5603.73 | 143.87%
Qwen--Qwen3-14B | bf16 | 2 | 16 | 6644.69 | 10388.69 | 156.35%
Qwen--Qwen3-14B | bf16 | 3 | 1 | 580.89 | 943.23 | 162.38%
Qwen--Qwen3-14B | bf16 | 3 | 2 | 973.85 | 1445.23 | 148.40%
Qwen--Qwen3-14B | bf16 | 3 | 4 | 1441.24 | 2395.75 | 166.23%
Qwen--Qwen3-14B | bf16 | 3 | 8 | 2734.81 | 4153.81 | 151.89%
Qwen--Qwen3-14B | bf16 | 3 | 16 | 4647.82 | 7021.95 | 151.08%
meta-llama--Llama-3.3-70B-Instruct | bf16 | 3 | 1 | 2810.45 | 5134.92 | 182.71%
meta-llama--Llama-3.3-70B-Instruct | bf16 | 3 | 2 | 4981.97 | 9189.88 | 184.46%
meta-llama--Llama-3.3-70B-Instruct | bf16 | 3 | 4 | 9462.15 | 15257.43 | 161.25%
meta-llama--Llama-3.3-70B-Instruct | bf16 | 3 | 8 | 14804.28 | 26549.99 | 179.34%
meta-llama--Llama-3.3-70B-Instruct | bf16 | 3 | 16 | 25490.63 | 45226.34 | 177.42%